start moving this out, the global should go
