+++ projects/libsodium/fake_random.h
@@ -1,0 +1,4 @@
+extern const unsigned char *SEED;
+extern size_t SEED_SIZE;
+
+struct randombytes_implementation fake_random_implementation;
